requirements.txt
argparse == 3.2
Pillow == 7.2.0
